pyproject.toml
[project]
name = "llm-test"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "Yun Lian", email = "bjliany@gmail.com"},
]
dependencies = [
    "aiohttp==3.9.3",
    "aiosignal==1.3.1",
    "alabaster==0.7.16",
    "annotated-types==0.6.0",
    "anyio==4.3.0",
    "appnope==0.1.4",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asgiref==3.7.2",
    "asttokens==2.4.1",
    "async-lru==2.0.4",
    "attrs==23.2.0",
    "Babel==2.14.0",
    "backoff==2.2.1",
    "bcrypt==4.1.2",
    "beautifulsoup4==4.12.3",
    "bleach==6.1.0",
    "bokeh==3.3.4",
    "Brotli==1.1.0",
    "build==1.1.1",
    "cachetools==5.3.3",
    "cattrs==23.2.3",
    "certifi==2024.2.2",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "chroma-hnswlib==0.7.3",
    "chromadb==0.4.24",
    "click==8.1.7",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "comm==0.2.1",
    "contourpy==1.2.0",
    "dataclasses-json==0.6.4",
    "debugpy==1.8.1",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "Deprecated==1.2.14",
    "distro==1.9.0",
    "docarray==0.40.0",
    "docutils==0.20.1",
    "esbonio==0.16.4",
    "executing==2.0.1",
    "faiss-cpu==1.8.0",
    "fastapi==0.110.0",
    "fastjsonschema==2.19.1",
    "filelock==3.13.1",
    "flatbuffers==24.3.7",
    "fqdn==1.5.1",
    "frozenlist==1.4.1",
    "fsspec==2024.3.0",
    "gitdb==4.0.11",
    "GitPython==3.1.42",
    "google-auth==2.28.2",
    "googleapis-common-protos==1.63.0",
    "grpcio==1.62.1",
    "h11==0.14.0",
    "httpcore==1.0.4",
    "httptools==0.6.1",
    "httpx==0.25.2",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.21.4",
    "humanfriendly==10.0",
    "idna==3.6",
    "imagesize==1.4.1",
    "importlib-metadata==6.11.0",
    "importlib-resources==6.3.0",
    "ipykernel==6.29.3",
    "ipython==8.22.2",
    "ipywidgets==8.1.2",
    "isoduration==20.11.0",
    "jedi==0.19.1",
    "Jinja2==3.1.3",
    "json5==0.9.22",
    "jsonpatch==1.33",
    "jsonpointer==2.4",
    "jsonschema==4.21.1",
    "jsonschema-specifications==2023.12.1",
    "jupyter==1.0.0",
    "jupyter-console==6.6.3",
    "jupyter-events==0.9.0",
    "jupyter-lsp==2.2.4",
    "jupyter-client==8.6.0",
    "jupyter-core==5.7.1",
    "jupyter-server==2.13.0",
    "jupyter-server-terminals==0.5.2",
    "jupyterlab==4.1.4",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.25.3",
    "jupyterlab-widgets==3.0.10",
    "kubernetes==29.0.0",
    "langchain==0.1.12",
    "langchain-cli==0.0.21",
    "langchain-community==0.0.28",
    "langchain-core==0.1.32",
    "langchain-experimental==0.0.54",
    "langchain-openai==0.0.8",
    "langchain-text-splitters==0.0.1",
    "langchainhub==0.1.15",
    "langserve==0.0.51",
    "langsmith==0.1.23",
    "linkify-it-py==2.0.3",
    "lsprotocol==2023.0.1",
    "Markdown==3.6",
    "markdown-it-py==3.0.0",
    "MarkupSafe==2.1.5",
    "marshmallow==3.21.1",
    "matplotlib-inline==0.1.6",
    "mdit-py-plugins==0.4.0",
    "mdurl==0.1.2",
    "mistune==3.0.2",
    "mmh3==4.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "multidict==6.0.5",
    "mutagen==1.47.0",
    "mypy-extensions==1.0.0",
    "nbclient==0.9.0",
    "nbconvert==7.16.2",
    "nbformat==5.9.2",
    "nest-asyncio==1.6.0",
    "notebook==7.1.1",
    "notebook-shim==0.2.4",
    "numexpr==2.9.0",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "ollama==0.1.7",
    "onnxruntime==1.17.1",
    "openai==1.13.3",
    "opentelemetry-api==1.23.0",
    "opentelemetry-exporter-otlp-proto-common==1.23.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.23.0",
    "opentelemetry-instrumentation==0.44b0",
    "opentelemetry-instrumentation-asgi==0.44b0",
    "opentelemetry-instrumentation-fastapi==0.44b0",
    "opentelemetry-proto==1.23.0",
    "opentelemetry-sdk==1.23.0",
    "opentelemetry-semantic-conventions==0.44b0",
    "opentelemetry-util-http==0.44b0",
    "orjson==3.9.15",
    "overrides==7.7.0",
    "packaging==23.2",
    "pandas==2.2.1",
    "pandocfilters==1.5.1",
    "panel==1.3.8",
    "param==2.0.2",
    "parso==0.8.3",
    "pexpect==4.9.0",
    "pillow==10.2.0",
    "platformdirs==4.2.0",
    "posthog==3.5.0",
    "prometheus-client==0.20.0",
    "prompt-toolkit==3.0.43",
    "protobuf==4.25.3",
    "psutil==5.9.8",
    "ptyprocess==0.7.0",
    "pulsar-client==3.4.0",
    "pure-eval==0.2.2",
    "pyasn1==0.5.1",
    "pyasn1-modules==0.3.0",
    "pycparser==2.21",
    "pycryptodomex==3.20.0",
    "pydantic==2.6.3",
    "pydantic-core==2.16.3",
    "pydub==0.25.1",
    "pygls==1.3.0",
    "Pygments==2.17.2",
    "pypdf==4.1.0",
    "PyPika==0.48.9",
    "pyproject-hooks==1.0.0",
    "pyspellchecker==0.8.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-json-logger==2.0.7",
    "pytz==2024.1",
    "pyviz-comms==3.0.1",
    "PyYAML==6.0.1",
    "pyzmq==25.1.2",
    "qtconsole==5.5.1",
    "QtPy==2.4.1",
    "redlines==0.4.2",
    "referencing==0.33.0",
    "regex==2023.12.25",
    "requests==2.31.0",
    "requests-oauthlib==1.4.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rich==13.7.1",
    "rich-click==1.7.4",
    "rpds-py==0.18.0",
    "rsa==4.9",
    "Send2Trash==1.8.2",
    "shellingham==1.5.4",
    "six==1.16.0",
    "smmap==5.0.1",
    "sniffio==1.3.1",
    "snowballstemmer==2.2.0",
    "soupsieve==2.5",
    "Sphinx==7.2.6",
    "sphinxcontrib-applehelp==1.0.8",
    "sphinxcontrib-devhelp==1.0.6",
    "sphinxcontrib-htmlhelp==2.0.5",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==1.0.7",
    "sphinxcontrib-serializinghtml==1.1.10",
    "SQLAlchemy==2.0.28",
    "sse-starlette==1.8.2",
    "stack-data==0.6.3",
    "starlette==0.36.3",
    "sympy==1.12",
    "tavily-python==0.3.1",
    "tenacity==8.2.3",
    "terminado==0.18.0",
    "tiktoken==0.5.2",
    "tinycss2==1.2.1",
    "tokenizers==0.15.2",
    "tomlkit==0.12.4",
    "tornado==6.4",
    "tqdm==4.66.2",
    "traitlets==5.14.1",
    "typer==0.9.0",
    "types-python-dateutil==2.8.19.20240311",
    "types-requests==2.31.0.20240311",
    "typing-inspect==0.9.0",
    "typing-extensions==4.10.0",
    "tzdata==2024.1",
    "uc-micro-py==1.0.3",
    "uri-template==1.3.0",
    "urllib3==2.2.1",
    "uvicorn==0.23.2",
    "uvloop==0.19.0",
    "watchfiles==0.21.0",
    "wcwidth==0.2.13",
    "webcolors==1.13",
    "webencodings==0.5.1",
    "websocket-client==1.7.0",
    "websockets==12.0",
    "widgetsnbextension==4.0.10",
    "wikipedia==1.4.0",
    "wrapt==1.16.0",
    "xyzservices==2023.10.1",
    "yarl==1.9.4",
    "yt-dlp==2024.3.10",
    "zipp==3.18.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "lian"}


[tool.pdm]
distribution = false
